backend/requirements.txt
fastapi
uvicorn
python-multipart
opencv-python-headless
numpy
scikit-learn
google-genai